Grafo H Correcion Matriz Adyacencias
Ya que ahora la matriz de adyacencias trabaja con valores decimales, la converti a double

+++ Grafo.h
@@ -5,7 +5,8 @@
 #include "Cola.h"
 
 class Grafo{
-    int **matAdj;
+    //int **matAdj;
+    double **matAdj;
     Vertice *ListaV;
     int numV;
     fstream arch;
